feat(threading): Refactored thread scheduling, signals, and synchronization
- Refactored existing code to use the new threading infrastructure.
- Removed deprecated syscalls and macros.
- Introduced new syscalls for thread control, signal management, synchronization, and scheduling policies.
- Updated thread synchronization primitives, including mutexes and semaphores.
- Updated helper functions for managing ready queues, sleep queues, and thread states.
- Updated example programs to demonstrate thread joining, sleeping, yielding, mutexes, semaphores, and signals (signals are still WIP).
- Added extensive tracing for debugging.
