+++ sys/SRCFILES
@@ -64,6 +64,11 @@ COBJS = \
 	proc_threads_signal.c \
 	proc_threads_policy.c \
 	proc_threads_sync.c \
+	proc_threads_queue.c \
+	proc_threads_helper.c \
+	proc_threads_scheduler.c \
+	proc_threads_sleep_yield.c \
+	proc_threads_syscall.c \
 	procfs.c \
 	profil.c \
 	ptrace.c \
