+++ sys/k_exec.c
@@ -282,7 +282,7 @@ sys_pexec(short mode, const void *p1, const void *p2, const void *p3)
 				void *arg = (void*) (unsigned long) p2;
 				void *stack = (void*) (unsigned long) p3;
 				
-				return sys_p_createthread(func, arg, stack);
+				return proc_thread_create(func, arg, stack);
 			}
 		default:
 		{
