+++ sys/proc_threads_policy.c
@@ -1,18 +1,12 @@
-#include "proc_threads.h"
+#include "proc_threads_policy.h"
+#include "proc_threads_queue.h"
+#include "proc_threads_helper.h"
+#include "proc_threads_scheduler.h"
 
-/* Thread scheduling */
-
-int set_thread_timeslice(struct thread *t, long timeslice);
-int set_thread_policy(struct thread *t, enum sched_policy policy, int priority);
-int get_rr_interval(struct thread *t, long *interval);
-
-/* Thread scheduling system calls */
-long _cdecl sys_p_sched_yield(void);
-long _cdecl sys_p_get_thread_timeslice(long tid, long *timeslice, long *remaining);
-long _cdecl sys_p_set_thread_timeslice(long tid, long timeslice);
-long _cdecl sys_p_getrrinterval(long tid, long *interval);
-long _cdecl sys_p_getschedparam(long tid, long *policy, long *priority);
-long _cdecl sys_p_setschedparam(long tid, long policy, long priority);
+/* Forward declarations for internal functions */
+static int set_thread_timeslice(struct thread *t, long timeslice);
+static int set_thread_policy(struct thread *t, enum sched_policy policy, int priority);
+static int get_rr_interval(struct thread *t, long *interval);
 
 /**
  * Get the time quantum for SCHED_RR threads
@@ -21,10 +15,10 @@ long _cdecl sys_p_setschedparam(long tid, long policy, long priority);
  * @param interval Pointer to store the time quantum in milliseconds
  * @return 0 on success, negative error code on failure
  */
-int get_rr_interval(struct thread *t, long *interval)
+static int get_rr_interval(struct thread *t, long *interval)
 {
     if (!t || !interval)
-        return -EINVAL;
+        return EINVAL;
         
     *interval = t->timeslice * 5; // Convert ticks to milliseconds (5ms per tick)
     return 0;
@@ -38,14 +32,14 @@ int get_rr_interval(struct thread *t, long *interval)
  * @param priority The new priority (0-99, with 0 being lowest)
  * @return 0 on success, negative error code on failure
  */
-int set_thread_policy(struct thread *t, enum sched_policy policy, int priority)
+static int set_thread_policy(struct thread *t, enum sched_policy policy, int priority)
 {
     if (!t || t->magic != CTXT_MAGIC)
-        return -EINVAL;
+        return EINVAL;
 
     // Validate process pointer
     if (!t->proc)
-        return -EINVAL;
+        return EINVAL;
         
     // Log the request
     TRACE_THREAD("POLICY: Setting thread %d policy to %d, priority to %d (current: policy=%d, pri=%d)",
@@ -56,7 +50,7 @@ int set_thread_policy(struct thread *t, enum sched_policy policy, int priority)
 
     // Validate policy and priority
     if (policy != SCHED_FIFO && policy != SCHED_RR && policy != SCHED_OTHER)
-        return -EINVAL;
+        return EINVAL;
 
     // SCHED_OTHER can only use priority 0
     if (policy == SCHED_OTHER && priority != 0)
@@ -130,7 +124,7 @@ int set_thread_policy(struct thread *t, enum sched_policy policy, int priority)
     // If this is the current thread and we lowered priority, trigger a reschedule
     if (t == t->proc->current_thread && priority < old_priority) {
         spl(sr);
-        schedule();
+        proc_thread_schedule();
         return 0;
     }
     
@@ -141,13 +135,13 @@ int set_thread_policy(struct thread *t, enum sched_policy policy, int priority)
 /**
  * System call to set thread scheduling policy and priority
  */
-long _cdecl sys_p_setschedparam(long tid, long policy, long priority)
+long proc_thread_set_schedparam(long tid, long policy, long priority)
 {
     struct proc *p = curproc;
     struct thread *t = NULL;
     
     if (!p)
-        return -EINVAL;
+        return EINVAL;
         
     // Find thread by ID
     if (tid < 0) {
@@ -161,7 +155,7 @@ long _cdecl sys_p_setschedparam(long tid, long policy, long priority)
         }
         
         if (!t)
-            return -ESRCH; // No such thread
+            return ESRCH; // No such thread
     }
     
     return set_thread_policy(t, (enum sched_policy)policy, (int)priority);
@@ -170,13 +164,13 @@ long _cdecl sys_p_setschedparam(long tid, long policy, long priority)
 /**
  * System call to get thread scheduling policy and priority
  */
-long _cdecl sys_p_getschedparam(long tid, long *policy, long *priority)
+long proc_thread_get_schedparam(long tid, long *policy, long *priority)
 {
     struct proc *p = curproc;
     struct thread *t = NULL;
     
     if (!p || !policy || !priority)
-        return -EINVAL;
+        return EINVAL;
         
     // Find thread by ID
     if (tid < 0) {
@@ -190,7 +184,7 @@ long _cdecl sys_p_getschedparam(long tid, long *policy, long *priority)
         }
         
         if (!t)
-            return -ESRCH; // No such thread
+            return ESRCH; // No such thread
     }
     
     *policy = t->policy;
@@ -202,13 +196,13 @@ long _cdecl sys_p_getschedparam(long tid, long *policy, long *priority)
 /**
  * System call to get the RR interval for a thread
  */
-long _cdecl sys_p_getrrinterval(long tid, long *interval)
+long proc_thread_get_rrinterval(long tid, long *interval)
 {
     struct proc *p = curproc;
     struct thread *t = NULL;
     
     if (!p || !interval)
-        return -EINVAL;
+        return EINVAL;
         
     // Find thread by ID
     if (tid < 0) {
@@ -229,90 +223,15 @@ long _cdecl sys_p_getrrinterval(long tid, long *interval)
 }
 
 /**
- * System call for thread to voluntarily yield the CPU
- */
-long _cdecl sys_p_sched_yield(void)
-{
-    struct proc *p = curproc;
-    struct thread *t;
-    
-    if (!p || !p->current_thread)
-        return -EINVAL;
-        
-    t = p->current_thread;
-    
-    // For SCHED_FIFO and SCHED_RR, move to end of same-priority list
-    if (t->policy == SCHED_FIFO || t->policy == SCHED_RR) {
-        unsigned short sr = splhigh();
-        
-        // Only if we're in RUNNING state
-        if (t->state == THREAD_STATE_RUNNING) {
-            // Change to READY and add to end of ready queue
-            atomic_thread_state_change(t, THREAD_STATE_READY);
-            
-            // // Remove from ready queue if present
-            // remove_from_ready_queue(t);
-            
-            // Add to ready queue (will be added at end of same-priority list)
-            add_to_ready_queue(t);
-            
-            // Force a reschedule
-            spl(sr);
-            schedule();
-            return 0;
-        }
-        
-        spl(sr);
-    } else {
-        // For SCHED_OTHER, just call schedule
-        schedule();
-    }
-    
-    return 0;
-}
-
-/**
- * Main entry point for thread scheduling system calls
- */
-long _cdecl sys_p_thread_sched(long func, long arg1, long arg2, long arg3)
-{
-    TRACE_THREAD("IN KERNEL: sys_p_thread_sched: func=%ld, arg1=%ld, arg2=%ld, arg3=%ld", 
-                func, arg1, arg2, arg3);
-    
-    switch (func) {
-        case PSCHED_SETPARAM:
-            return sys_p_setschedparam(arg1, arg2, arg3);
-            
-        case PSCHED_GETPARAM:
-            return sys_p_getschedparam(arg1, (long*)arg2, (long*)arg3);
-            
-        case PSCHED_YIELD:
-            return sys_p_sched_yield();
-            
-        case PSCHED_GETRRINTERVAL:
-            return sys_p_getrrinterval(arg1, (long*)arg2);
-            
-        case PSCHED_SET_TIMESLICE:
-            return sys_p_set_thread_timeslice(arg1, arg2);
-            
-        case PSCHED_GET_TIMESLICE:
-            return sys_p_get_thread_timeslice(arg1, (long*)arg2, (long*)arg3);
-            
-        default:
-            return -EINVAL;
-    }
-}
-
-/**
  * System call to set thread timeslice
  */
-long _cdecl sys_p_set_thread_timeslice(long tid, long timeslice)
+long proc_thread_set_timeslice(long tid, long timeslice)
 {
     struct proc *p = curproc;
     struct thread *t = NULL;
     
     if (!p)
-        return -EINVAL;
+        return EINVAL;
         
     // Find thread by ID
     if (tid < 0) {
@@ -326,7 +245,7 @@ long _cdecl sys_p_set_thread_timeslice(long tid, long timeslice)
         }
         
         if (!t)
-            return -ESRCH; // No such thread
+            return ESRCH; // No such thread
     }
     
     return set_thread_timeslice(t, timeslice);
@@ -335,13 +254,13 @@ long _cdecl sys_p_set_thread_timeslice(long tid, long timeslice)
 /**
  * System call to get thread timeslice
  */
-long _cdecl sys_p_get_thread_timeslice(long tid, long *timeslice, long *remaining)
+long proc_thread_get_timeslice(long tid, long *timeslice, long *remaining)
 {
     struct proc *p = curproc;
     struct thread *t = NULL;
     
     if (!p || !timeslice)
-        return -EINVAL;
+        return EINVAL;
         
     // Find thread by ID
     if (tid < 0) {
@@ -372,14 +291,14 @@ long _cdecl sys_p_get_thread_timeslice(long tid, long *timeslice, long *remainin
  * @param timeslice The new timeslice in ticks
  * @return 0 on success, negative error code on failure
  */
-int set_thread_timeslice(struct thread *t, long timeslice)
+static int set_thread_timeslice(struct thread *t, long timeslice)
 {
     if (!t || t->magic != CTXT_MAGIC)
-        return -EINVAL;
+        return EINVAL;
 
     // Validate process pointer
     if (!t->proc)
-        return -EINVAL;
+        return EINVAL;
         
     // Enforce minimum timeslice        
     if (timeslice < t->proc->thread_min_timeslice)
@@ -436,12 +355,12 @@ void update_thread_timeslice(struct thread *t)
 }
 
 // Function to set thread scheduling policy
-long _cdecl sys_p_setthreadpolicy(enum sched_policy policy, short priority, short timeslice)
+long proc_thread_set_policy(enum sched_policy policy, short priority, short timeslice)
 {
     struct proc *p = curproc;
     if (!p || !p->current_thread) {
-        TRACE_THREAD("sys_p_setthreadpolicy: invalid current process/thread");
-        return -EINVAL;
+        TRACE_THREAD("proc_thread_set_policy: invalid current process/thread");
+        return EINVAL;
     }
 
     struct thread *current = p->current_thread;
