+++ sys/signal.c
@@ -29,7 +29,7 @@
 # include "proc_help.h"
 # include "util.h"
 
-extern int thread_aware_raise(struct proc *p, int sig);
+extern int proc_thread_signal_aware_raise(struct proc *p, int sig);
 
 /* send_sig: Send signal SIG to process P. If PRIV is non-zero then
  * the signal is generated by the kernel and no permissions should
@@ -209,7 +209,7 @@ post_sig (PROC *p, ushort sig)
 	/* Check for thread-specific signal handling */
 	if (p->p_sigacts && p->p_sigacts->thread_signals) {
 		/* Try thread-aware signal delivery */
-		if (thread_aware_raise(p, sig) == 0) {
+		if (proc_thread_signal_aware_raise(p, sig) == 0) {
 			delivered = 1;
 		}
 	}
