+++ sys/syscall_vectors.c
@@ -521,17 +521,17 @@ Func dos_tab [DOS_MAX] =
 	/* 0x182 */	(Func)	sys_f_opendir,	/* 1.17 */
 	/* 0x183 */		sys_f_dirfd,	/* 1.17 */
 	/* 0x184 */		sys_enosys,		/* reserved */
-	/* 0x185 */		sys_p_thread_sched,
-	/* 0x186 */		sys_p_sleepthread,		/* reserved */
+	/* 0x185 */		sys_p_thread_sched_policy,
+	/* 0x186 */		sys_enosys,		/* reserved */
 	/* 0x187 */		sys_enosys,		/* reserved */
 	/* 0x188 */		sys_enosys,		/* reserved */
 	/* 0x189 */		sys_enosys,		/* reserved */
-	/* 0x18a */		sys_p_exitthread,
+	/* 0x18a */		sys_p_thread_ctrl,
 	/* 0x18b */		sys_enosys,		/* reserved */
 	/* 0x18c */		sys_enosys,		/* reserved */
-	/* 0x18d */		sys_p_threadop,
-	/* 0x18e */		sys_p_threadsignal,
-	/* 0x18f */		sys_p_thread_alarm,
+	/* 0x18d */		sys_p_thread_sync,
+	/* 0x18e */		sys_p_thread_signal,
+	/* 0x18f */		sys_enosys,		/* reserved */
 
 	/* 0x190 */		/* DOS_MAX */
 };
