+++ sys/syscalls.master
@@ -520,16 +520,16 @@ opcode	ret	class name	args */
 0x182		Ffdopendir	(short fd) /* since 1.17 */
 0x183		Fdirfd		(long handle) /* since 1.17 */
 0x184		undefined
-0x185	void	Pthread_sched	(long func, long arg1, long arg2, long arg3)
-0x186	long	Psleepthread	(long ms)
+0x185		undefined
+0x186		undefined
 0x187		undefined
 0x188		undefined
 0x189		undefined
-0x18a	long	Pexitthread	(long mode, long arg1, long arg2)
+0x18a		undefined
 0x18b		undefined
 0x18c		undefined
-0x18d	long	Pthreadop	(long operator, void *arg)
-0x18e	long	Pthreadsignal	(long func, long arg1, long arg2)
+0x18d		undefined
+0x18e		undefined
 0x18f		undefined
 
 0x190	MAX
